backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
asyncpg==0.29.0

# Settings management
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0

# WebSocket support (included in uvicorn[standard])
websockets==12.0

# Security
python-multipart==0.0.6
